adding submodules and update setup flow

+++ setup.py
@@ -39,31 +39,40 @@ __copyright__ = "Copyright 2017, Xilinx"
 __email__ = "yunq@xilinx.com"
 
 
+GIT_DIR = os.path.dirname(os.path.realpath(__file__))
+
 # Update boot partition
 def update_boot():
     boot_mount = '/mnt/boot'
-    subprocess.check_call(['mkdir', boot_mount])
-    subprocess.check_call(['mount', '/dev/mmcblk0p1', boot_mount])
+    if not os.path.exists('/mnt/boot'):
+        subprocess.check_call(['mkdir', boot_mount])
+        try:
+            subprocess.check_call(['mount', '/dev/mmcblk0p1', boot_mount])
+        finally:
+            subprocess.check_call(['umount', '/dev/mmcblk0p1'])
 
     backup_folder = boot_mount + '/BOOT_PARTITION_{}'.format(
         datetime.now().strftime("%Y_%m_%d_%H_%M_%S"))
     subprocess.check_call(['mkdir', backup_folder])
-    boot_file = ['BOOT.BIN', 'devicetree.dtb', 'uEnv.txt', 'uImage']
+    boot_file = ['BOOT.bin', 'devicetree.dtb', 'uEnv.txt', 'uImage']
     for file in boot_file:
-        shutil.copy2(boot_mount + '/' + file,
-                     backup_folder + '/')
+        if os.path.isfile(boot_mount + '/' + file):
+            shutil.copy2(boot_mount + '/' + file,
+                         backup_folder + '/')
     for file in boot_file:
-        shutil.copy2('boot_files/' + file,
+        shutil.copy2(GIT_DIR + '/boot_files/' + file,
                      boot_mount + '/')
+    print("Update boot files done ...")
 
 
 # Install packages
 def install_packages():
-    subprocess.check_call(['apt-get', 'install',
+    subprocess.check_call(['apt-get', '--yes', '--force-yes', 'install',
                            'tcpdump', 'iptables', 'ebtables', 'bridge-utils'])
     subprocess.check_call(['pip3.6', 'install',
                            'scapy-python3', 'wurlitzer',
                            'pytest-runner', 'paho-mqtt'])
+    print("Installing packages done ...")
 
 
 # Update interfaces
@@ -72,49 +81,51 @@ def update_interfaces():
     backup_file = eth0_file + '_{}'.format(
         datetime.now().strftime("%Y_%m_%d_%H_%M_%S"))
     shutil.copy2(eth0_file, backup_file)
-    shutil.copy2('interfaces.d/eth0', eth0_file)
+    shutil.copy2(GIT_DIR + '/interfaces.d/eth0', eth0_file)
+    print("Update interface files done ...")
 
 
 # Build submodules
-def build_submodules(submodule_path):
-    if os.path.exists(submodule_path + '/.git'):
-        subprocess.check_call(['git', 'submodule', 'init'])
-        subprocess.check_call(['git', 'submodule', 'update'])
+def build_submodules():
+    subprocess.check_call(['git', 'submodule', 'init'])
+    subprocess.check_call(['git', 'submodule', 'update'])
+    print("Update submodules done ...")
 
 
 # Run makefiles
 def run_make(src_path, output_lib):
-    status = subprocess.check_call(["make", "-C", src_path])
+    status = subprocess.check_call(["make", "-C", GIT_DIR + '/' + src_path])
     if status is not 0:
         print("Error while running make for", output_lib, "Exiting..")
         sys.exit(1)
+    print("Running make for " + output_lib + " done ...")
 
 
 # Notebook delivery
 def fill_notebooks():
-    src_nb = 'notebooks/*'
+    src_nb = GIT_DIR + '/notebooks'
     dst_nb_dir = '/home/xilinx/jupyter_notebooks/networking'
-    subprocess.check_call(['cp', '-rf', src_nb, dst_nb_dir])
+    shutil.copytree(src_nb, dst_nb_dir)
 
-    if os.path.exists(new_nb_dir):
-        for folder in ['/kernel_module', '/broker_client']:
-            subprocess.check_call(['cp', '-rf', folder, dst_nb_dir])
+    for folder in ['/kernel_module', '/broker_client']:
+        shutil.copytree(GIT_DIR + folder, dst_nb_dir + folder)
+    print("Filling notebooks done ...")
 
 
 # Move overlay
 def fill_overlay():
-    src_overlay = 'overlays/*'
+    src_overlay = GIT_DIR + '/overlays/mqttsn'
     dst_overlay_dir = '/home/xilinx/pynq/overlays/'
     subprocess.check_call(['cp', '-rf', src_overlay, dst_overlay_dir])
+    print("Filling overlays done ...")
 
 
 if len(sys.argv) > 1 and sys.argv[1] == 'install':
     update_boot()
     install_packages()
     update_interfaces()
-    build_submodules("org.eclipse.mosquitto.rsmb")
-    build_submodules("mqtt-sn-tools")
-    run_make("org.eclipse.mosquitto.rsmb/rsmb/src/", "broker_mqtts")
+    build_submodules()
+    run_make("rsmb/rsmb/src/", "broker_mqtts")
     fill_notebooks()
     fill_overlay()
     print("Please reboot the board to finish the setup.")
@@ -129,6 +140,7 @@ setup(name='pynq_networking',
       packages=find_packages(),
       download_url='https://github.com/Xilinx/PYNQ-Networking',
       package_data={
-          '': ['tests/*', 'js/*', '*.bin', '*.so', '*.bit', '*.tcl', '*.pdm'],
+          '': ['tests/*', '*.ipynb', '*.bin', 
+               '*.ko', '*.so', '*.bit', '*.tcl'],
       }
       )
